feat: 添加 AttachToExpectedRegexAction 以合并源节点关键词并覆盖目标节点的预期正则表达式，同时删除不再使用的 CreditShopping 相关文件
