+++ agent/go-service/common/attachregex/register.go
@@ -1,0 +1,7 @@
+package attachregex
+
+import maa "github.com/MaaXYZ/maa-framework-go/v4"
+
+func Register() {
+	maa.AgentServerRegisterCustomAction("AttachToExpectedRegexAction", &AttachToExpectedRegexAction{})
+}
