--- agent/go-service/creditshopping/register.go
@@ -1,8 +1,0 @@
-package creditshopping
-
-import maa "github.com/MaaXYZ/maa-framework-go/v4"
-
-// Register registers all custom action components for creditshopping package
-func Register() {
-	maa.AgentServerRegisterCustomAction("CreditShoppingParseParams", &CreditShoppingParseParams{})
-}
