+++ agent/go-service/register.go
@@ -8,11 +8,11 @@ import (
 	"github.com/MaaXYZ/MaaEnd/agent/go-service/batchaddfriends"
 	"github.com/MaaXYZ/MaaEnd/agent/go-service/bettersliding"
 	"github.com/MaaXYZ/MaaEnd/agent/go-service/blueprintimport"
+	"github.com/MaaXYZ/MaaEnd/agent/go-service/common/attachregex"
 	"github.com/MaaXYZ/MaaEnd/agent/go-service/common/autoaltclick"
 	"github.com/MaaXYZ/MaaEnd/agent/go-service/common/charactercontroller"
 	"github.com/MaaXYZ/MaaEnd/agent/go-service/common/clearhitcount"
 	"github.com/MaaXYZ/MaaEnd/agent/go-service/common/subtask"
-	"github.com/MaaXYZ/MaaEnd/agent/go-service/creditshopping"
 	"github.com/MaaXYZ/MaaEnd/agent/go-service/dailyrewards"
 	"github.com/MaaXYZ/MaaEnd/agent/go-service/essencefilter"
 	"github.com/MaaXYZ/MaaEnd/agent/go-service/expressionrecognition"
@@ -43,6 +43,7 @@ func registerAll() {
 	subtask.Register()
 	clearhitcount.Register()
 	expressionrecognition.Register()
+	attachregex.Register()
 	autoaltclick.Register()
 	charactercontroller.Register()
 
@@ -53,7 +54,6 @@ func registerAll() {
 	bettersliding.Register()
 	essencefilter.Register()
 	dailyrewards.Register()
-	creditshopping.Register()
 	maptracker.Register()
 	batchaddfriends.Register()
 	autoecofarm.Register()
