+++ docs/en_us/developers/custom.md
@@ -50,6 +50,42 @@ Example file: [`SubTask.json`](../../../assets/resource/pipeline/Interface/Examp
 
 Example file: [`ClearHitCount.json`](../../../assets/resource/pipeline/Interface/Example/ClearHitCount.json)
 
+### AttachToExpectedRegexAction
+
+`AttachToExpectedRegexAction` is implemented in `agent/go-service/common/attachregex`. It generically reads keywords from `attach` of source nodes, then writes the merged whitelist regex into one target OCR node's `expected`.
+
+- Parameters:
+    - `target: string`: required target node name whose `expected` will be overridden.
+    - `keys: string[]`: required source node names to traverse for `attach`.
+
+Behavior:
+
+- `attach` values support `string` and `string[]`; values are trimmed, deduplicated, and regex-escaped.
+- If the keyword list is empty, it generates `a^` (never matches).
+- The final result is applied through `OverridePipeline` to the target node's `expected`.
+
+Example:
+
+```json
+{
+    "action": "Custom",
+    "custom_action": "AttachToExpectedRegexAction",
+    "custom_action_param": {
+        "target": "Priority2OCR",
+        "keys": [
+            "Priority2OCR",
+            "Priority2OCR_CanNotAfford"
+        ]
+    }
+}
+```
+
+Compatibility note:
+
+- Credit shop has switched to direct use of `AttachToExpectedRegexAction`.
+- If multiple targets need override, prefer multiple `Custom` nodes chained by `next` in Pipeline.
+- Other tasks should also prefer this generic action name to avoid business coupling.
+
 ---
 
 ## Custom Recognition
